Consolidate blinking for wind into the existing script and activate it using a variable inside the metar.py script
